requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-bootstrap4==22.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
drf-yasg==1.21.3
easy-thumbnails==2.8.3
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
serializers==0.2.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
